% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogrampn.R
\name{bias.pn}
\alias{bias.pn}
\title{Unbiased estimator for  alpha (PN distribution)}
\usage{
 bias.pn(x, model)
}
\arguments{
\item{x}{the response vector}

\item{model}{a variable returned by \code{\link{pn.mle}}}
}

\description{
Unbiased estimator for \code{alpha} of PN distribution
}

\references{

Gupta RD, Gupta RC. (1998). Analyzing skewed data by power normal model.
\emph{Test}, \bold{17}, 197--210.

}
