% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitypn.R
\name{pn.dens}
\alias{pn.dens}
\title{Estimated densities (PN distribution)}
\usage{
pn.dens(x, model, log=FALSE, ylab=NULL, xlab = NULL, main = NULL, ...)
}
\arguments{
\item{x}{the response vector}

\item{model}{a variable returned by \code{\link{pn.mle}}}

\item{log}{Logical, plot log-density if TRUE (default = FALSE)}

\item{ylab}{Title of the ylab, if NULL default is selected}

\item{xlab}{Title of the xlab, if NULL default is selected}

\item{main}{Main Title, if NULL default is selected}

\item{...}{further arguments to \cite{\link{plot}}}
}
\description{
Plot the estimated density or log-density (PN)
}

