% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogrampn.R
\name{pn.hist}
\alias{pn.hist}
\title{Histrogram and estimated densities plots (PN distribution)}
\usage{
pn.hist(x, model, breaks, main,..., col.lines, lwd, lty  )
}
\arguments{
\item{x}{the response vector}

\item{model}{a variable returned by \code{\link{pn.mle}}}

\item{breaks}{the same option in \cite{\link{histogram}}}

\item{main}{the main title (have useful default values)}

\item{...}{further arguments to \cite{\link{histogram}}}

\item{col.lines}{line color}

\item{lwd}{line width}

\item{lty}{line type}
}
\description{
Plot the histogram along with the estimated density (PN)
}

