% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linespn.R
\name{pn.lines}
\alias{pn.lines}
\title{Plot lines of PN densities}
\usage{
pn.lines(x, model, log=FALSE, ...)
}
\arguments{
\item{x}{the response vector}

\item{model}{a variable returned by \code{\link{pn.mle}}}

\item{log}{Logical, plot log-density if TRUE (default = FALSE)}

\item{...}{further arguments to \code{\link{lines}}}
}
\description{
Add lines of PN estimated denisty or log-density in pn.dens or pn.hist plots.
}

