% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatepn.R
\name{pn.mle}
\alias{pn.mle}
\title{Fit univariate PN distribution (Classic)}
\usage{
pn.mle(x)
}
\arguments{
\item{x}{the response vector}
}

\description{
Return the estimative of power parameter for PN distribution
}



\references{

Gupta RD, Gupta RC. (1998). Analyzing skewed data by power normal model.
\emph{Test}, \bold{17}, 197--210.

}
