% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnd_functions_rp.R
\name{pnd_functions_rp}
\alias{pnd_functions_rp}
\alias{dpnd_rp}
\alias{ppnd_rp}
\alias{qpnd_rp}
\alias{rpnd_rp}
\title{The Power Normal Distribution with Reparametrization}
\usage{
dpnd_rp(x, lambda, xi, tau, log = FALSE)

ppnd_rp(x, lambda, xi, tau)

qpnd_rp(p, lambda, xi, tau)

rpnd_rp(n, lambda, xi, tau)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{lambda}{Numeric scalar: shape parameter.}

\item{xi}{Numeric scalar: reparametrized location parameter (i.e.,
the median).}

\item{tau}{Numeric scalar: reparametrized scale parameter
(i.e., interquantile range divided by median).}

\item{log}{Logical; if \code{TRUE}, log-density is returned.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
\code{dpnd_rp()} returns the PDF,
\code{ppnd_rp()} returns the CDF,
\code{qpnd_rp()} returns the quantile function, and
\code{rpnd_rp()} generates random values from the distribution.
}
\description{
Probability density function (PDF), cumulative distribution function (CDF),
quantile function, and random number generation for the power normal
distribution (PND) with the reparametrization proposed by Maruo et al.
(2017).
}
\details{
These functions are wrapper functions that calculate the original
parameters \eqn{\mu} and \eqn{\sigma} from the reparametrized parameters
\eqn{\xi_{0.5}} and \eqn{\tau}, using the \code{pnd_reparm} function.
The resulting values are passed to the underlying functions:
\code{dpnd}, \code{ppnd}, \code{qpnd}, and \code{rpnd}.
}
\examples{
dpnd_rp(x = 50, lambda = 0.2, xi = 100, tau = 0.8)
ppnd_rp(x = 50, lambda = 0.2, xi = 100, tau = 0.8)
qpnd_rp(p = 0.4, lambda = 0.2, xi = 100, tau = 0.8)
rpnd_rp(n = 10, lambda = 0.2, xi = 100, tau = 0.8)

}
\references{
Goto, M., & Inoue, T. (1980). Some properties of the power normal distribution.
\emph{Japanese Journal of Biometrics}, 1, 28–54.
\url{https://doi.org/10.5691/jjb.1.28}

Maruo, K., Yamabe, T., & Yamaguchi, Y. (2017). Statistical simulation based on
right-skewed distributions. \emph{Computational Statistics}, 32(3), 889–907.
\url{https://doi.org/10.1007/s00180-016-0664-4}
}
\seealso{
\code{\link[powerNormal]{pnd_reparm}}
}
