% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogrampn.R
\name{pn.bias}
\alias{pn.bias}
\title{Unbiased estimator for  alpha (PN distribution)}
\usage{
 pn.bias(x)
}
\arguments{
\item{x}{the response vector}

}

\description{
Unbiased estimator for \code{alpha} of PN distribution
}

\references{

Gupta RD, Gupta RC. (1998). Analyzing skewed data by power normal model.
\emph{Test}, \bold{17}, 197--210.

}
