% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogrampn.R
\name{IC.pn}
\alias{IC.pn}
\title{Confidence interval for alpha (PN distribution)}
\usage{
 IC.pn(x, model, p)
}
\arguments{
\item{x}{the response vector}

\item{model}{a variable returned by \code{\link{pn.mle}}}

\item{p}{confidence level}
}

\description{
Confidence interval for the power parameter of PN distribution
}


\references{

Gupta RD, Gupta RC. (1998). Analyzing skewed data by power normal model.
\emph{Test}, \bold{17}, 197--210.

}
